requirements.txt
torch==2.5.0   
torchvision==0.20.0
Pillow==10.1.0       
matplotlib==3.8.3